Finish first version of liquidity token creation
