+++ .env.development
@@ -51,5 +51,5 @@ REACT_APP_NODE_3 = "https://data-seed-prebsc-1-s2.binance.org:8545/"
 
 REACT_APP_SUMMITCHECK_API = "https://summtokenanalyzerapi.herokuapp.com"
 
-REACT_APP_CREATE_TOKEN_ADDRESS = "0xeD83E23D7Bd5472f389FD53DD627140d24dA509D"
+REACT_APP_CREATE_TOKEN_ADDRESS = "0x2796820Ab7bbB88200E4e1688606eaCfEe37e014"
 REACT_APP_CREATE_TOKEN_FEE_RECEIVER_ADDRESS = "0x1dC525231e5aC146B219f4eff5d7Ff485DcD81ed"
